♻️ Migrate the test into correct test suite.
